+++ app/src/main/java/com/hiroshi/cimoc/ui/custom/Option.java
@@ -55,4 +55,3 @@ public class Option extends FrameLayout {
     }
 
 }
-
